pyproject.toml
[project]
name = "debug-assist"
version = "1.0.0"
description = "Claude Code skill for systematic debugging assistance"
requires-python = ">=3.10"
license = {text = "MIT"}

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-timeout>=2.3",
    "python-frontmatter>=1.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: LLM-based tests requiring claude CLI and ANTHROPIC_API_KEY",
]
